+++ services/transitmux/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod api;
 mod error;
 pub mod otp;
 pub mod util;
